requirements.txt
asgiref==3.11.1
dj-database-url==3.1.2
Django==5.2.11
djangorestframework==3.16.1
gunicorn==25.1.0
packaging==26.0
pillow==12.1.1
psycopg2-binary==2.9.11
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
whitenoise==6.12.0
